review, gas optimizations and cleanup (#8)
* forge fmt

* move interfaces

* add fmt intructions

* small cleanup for readability

* add todo for toHex

* rm some comments

* seperation of concerns: models

* change max int val

* change tohex function

* rename and move TapeFixedFee

* Formatting and seperation of concerns (#7)

* forge fmt

* move interfaces

* add fmt intructions

* seperation of concerns: models

* remove loop

* optimise bondutils loop

* add disclaimer

* remove unnecessary addition

* remove unnecessary addition

* amout to amount typo

* clean up constructor

* remove dead code

* fix withdraw function

* clean up comments: TapeBondUtils

* add comments and clean dead comments Tape

* use via ir during build step

* undo via-ir in makefile

* undo removing of loop

* slither support (not working)

* working slither support using via ir

* mitigate reentrancy on buy functions (not full fix)

* mitigate reentrancy on buy functions

* remove setCartridgeParams and validate

* remove unnecessary funcs and test currencytoken

* add tests

* forge install: forge-std

v1.9.3

* add std

* add fixes from cart to tape

* remove gas reports

* revert getCurrencyAmoutToMintTokens to lyno implementation

* fix uri situation

* remove dead comments

+++ script/DeployCartridge.s.sol
@@ -1,22 +1,21 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity >=0.8.24;
 
-import { Script,console } from "forge-std/src/Script.sol";
+import {Script, console} from "forge-std/src/Script.sol";
 
-import { Create2 } from "@openzeppelin/contracts/utils/Create2.sol";
+import {Create2} from "@openzeppelin/contracts/utils/Create2.sol";
 
-
-import { CartridgeFeeModel } from "../src/CartridgeFeeModel.sol";
-import { CartridgeModel} from "../src/CartridgeModel.sol";
-import { CartridgeOwnershipModelWithProxy as OwnershipModel } from "../src/CartridgeOwnershipModelWithProxy.sol";
-import { BondingCurveModel } from "../src/BondingCurveModel.sol";
-import { CartridgeBondUtils } from "../src/CartridgeBondUtils.sol";
-import { Cartridge } from "../src/Cartridge.sol";
-
+import {CartridgeFeeModel} from "@models/CartridgeFeeModel.sol";
+import {CartridgeModel} from "@models/CartridgeModel.sol";
+import {CartridgeOwnershipModelWithProxy as OwnershipModel} from "@models/CartridgeOwnershipModelWithProxy.sol";
+import {BondingCurveModel} from "@models/BondingCurveModel.sol";
+import {CartridgeBondUtils} from "../src/CartridgeBondUtils.sol";
+import {Cartridge} from "../src/Cartridge.sol";
 
 contract DeployCartridge is Script {
     address constant DEPLOY_FACTORY = 0x4e59b44847b379578588920cA78FbF26c0B4956C;
     bytes32 constant SALT = bytes32(0);
+
     function run() external {
         uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
         address operatorAddress = vm.envAddress("OPERATOR_ADDRESS");
@@ -25,9 +24,9 @@ contract DeployCartridge is Script {
 
         console.logString("Deploying Cartridge Contracts");
 
-        // Cartridge Fee Model 
+        // Cartridge Fee Model
         bytes memory feeModelCode = abi.encodePacked(type(CartridgeFeeModel).creationCode);
-        address feeModelAddress = Create2.computeAddress(SALT, keccak256(feeModelCode),DEPLOY_FACTORY);
+        address feeModelAddress = Create2.computeAddress(SALT, keccak256(feeModelCode), DEPLOY_FACTORY);
         console.logString("Expected feeModelAddress");
         console.logAddress(feeModelAddress);
         if (checkSize(feeModelAddress) == 0) {
@@ -37,10 +36,10 @@ contract DeployCartridge is Script {
         } else {
             console.logString("Already deployed feeModelAddress");
         }
-        
-        // Cartridge Model 
+
+        // Cartridge Model
         bytes memory cartridgeModelCode = abi.encodePacked(type(CartridgeModel).creationCode);
-        address cartridgeModelAddress = Create2.computeAddress(SALT, keccak256(cartridgeModelCode),DEPLOY_FACTORY);
+        address cartridgeModelAddress = Create2.computeAddress(SALT, keccak256(cartridgeModelCode), DEPLOY_FACTORY);
         console.logString("Expected cartridgeModelAddress");
         console.logAddress(cartridgeModelAddress);
         if (checkSize(cartridgeModelAddress) == 0) {
@@ -50,10 +49,11 @@ contract DeployCartridge is Script {
         } else {
             console.logString("Already deployed cartridgeModelAddress");
         }
-        
-        // Ownership Model 
-        bytes memory ownershipModelCode = abi.encodePacked(type(OwnershipModel).creationCode,abi.encode(operatorAddress));
-        address ownershipModelAddress = Create2.computeAddress(SALT, keccak256(ownershipModelCode),DEPLOY_FACTORY);
+
+        // Ownership Model
+        bytes memory ownershipModelCode =
+            abi.encodePacked(type(OwnershipModel).creationCode, abi.encode(operatorAddress));
+        address ownershipModelAddress = Create2.computeAddress(SALT, keccak256(ownershipModelCode), DEPLOY_FACTORY);
         console.logString("Expected ownershipModelAddress");
         console.logAddress(ownershipModelAddress);
         if (checkSize(ownershipModelAddress) == 0) {
@@ -64,10 +64,9 @@ contract DeployCartridge is Script {
             console.logString("Already deployed ownershipModelAddress");
         }
 
-        
-        // Bonding Curve Model 
+        // Bonding Curve Model
         bytes memory bcModelCode = abi.encodePacked(type(BondingCurveModel).creationCode);
-        address bcModelAddress = Create2.computeAddress(SALT, keccak256(bcModelCode),DEPLOY_FACTORY);
+        address bcModelAddress = Create2.computeAddress(SALT, keccak256(bcModelCode), DEPLOY_FACTORY);
         console.logString("Expected bcModelAddress");
         console.logAddress(bcModelAddress);
         if (checkSize(bcModelAddress) == 0) {
@@ -77,10 +76,11 @@ contract DeployCartridge is Script {
         } else {
             console.logString("Already deployed bcModelAddress");
         }
-        
+
         // Cartridge Bond Utils
         bytes memory cartridgeBondUtilsCode = abi.encodePacked(type(CartridgeBondUtils).creationCode);
-        address cartridgeBondUtilsAddress = Create2.computeAddress(SALT, keccak256(cartridgeBondUtilsCode),DEPLOY_FACTORY);
+        address cartridgeBondUtilsAddress =
+            Create2.computeAddress(SALT, keccak256(cartridgeBondUtilsCode), DEPLOY_FACTORY);
         console.logString("Expected cartridgeBondUtilsAddress");
         console.logAddress(cartridgeBondUtilsAddress);
         if (checkSize(cartridgeBondUtilsAddress) == 0) {
@@ -90,16 +90,17 @@ contract DeployCartridge is Script {
         } else {
             console.logString("Already deployed cartridgeBondUtilsAddress");
         }
-        
+
         // Cartridge
-        bytes memory cartridgeCode = abi.encodePacked(type(Cartridge).creationCode,
+        bytes memory cartridgeCode = abi.encodePacked(
+            type(Cartridge).creationCode,
             abi.encode(
                 operatorAddress,
                 cartridgeBondUtilsAddress,
                 100 // max steps
             )
         );
-        address cartridgeAddress = Create2.computeAddress(SALT, keccak256(cartridgeCode),DEPLOY_FACTORY);
+        address cartridgeAddress = Create2.computeAddress(SALT, keccak256(cartridgeCode), DEPLOY_FACTORY);
         console.logString("Expected cartridgeAddress");
         console.logAddress(cartridgeAddress);
         if (checkSize(cartridgeAddress) == 0) {
@@ -117,12 +118,11 @@ contract DeployCartridge is Script {
         vm.stopBroadcast();
     }
 
-    function checkSize(address addr) public view returns(uint extSize) {
+    function checkSize(address addr) public view returns (uint256 extSize) {
         assembly {
             extSize := extcodesize(addr) // returns 0 if EOA, >0 if smart contract
         }
     }
-    
 }
 
 // # cartridge asset
@@ -133,4 +133,3 @@ contract DeployCartridge is Script {
 // COEFS="[1000000000000000,1000000000000000,2000000000000000]"
 
 // ARGS="$OPERATOR $MAX_STEPS $CURRENCY_TOKEN $TAPE_FEE_MODEL $TAPE_MODEL $OWNERSHIP_MODEL $BC_MODEL $TAPE_BOND_UTILS $MAX_SUPPLY $RANGES $COEFS"
-
